requirements.txt
Django>=5.0.0
torch>=2.0.0
torchvision>=0.15.0
pennylane>=0.30.0
pennylane-qiskit>=0.30.0
qiskit>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
requests>=2.31.0
tqdm>=4.65.0
asgiref>=3.7.0
sqlparse>=0.4.0
typing-extensions>=4.5.0
whitenoise>=6.5.0
gunicorn>=21.2.0
